// LondonFhirService.Manage.Client/src/models/views/audits/AuditDetailView.ts
export type AuditDetailView = {
    id: string;
    title: string;
    auditTypeText: string;
    message: string;
    logLevelText: string;
    logLevelClassName: string;
    correlationIdText: string;
    fileNameText: string;
    createdByText: string;
    createdDateText: string;
    updatedByText: string;
    updatedDateText: string;
    detailUrl: string;
};